% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{check.cocoa.input}
\alias{check.cocoa.input}
\title{Check CoCoA input}
\usage{
check.cocoa.input(
  mtx.data,
  celltype,
  cell2indv,
  indv2exp = NULL,
  celltype.mat = NULL
)
}
\arguments{
\item{mtx.data}{fileset.header ($mtx, $row, $col, $idx)}

\item{celltype}{celltype/cluster assignments (1 x cells or just a single string)}

\item{cell2indv}{cell-level individual assignments (cells x 2), cell -> indv}

\item{indv2exp}{individual treatment/exposure assignments (indiv x 2), indv -> exposure}

\item{celltype.mat}{celltype/cluster assignment matrix (cells x cell types)}
}
\description{
Check CoCoA input
}
