% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{fileset.list}
\alias{fileset.list}
\title{Create a list of MTX-related files}
\usage{
fileset.list(.hdr)
}
\arguments{
\item{.hdr}{file set header name}
}
\value{
a list of file names
}
\description{
Create a list of MTX-related files
}
