% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{make.sc.deg}
\alias{make.sc.deg}
\title{Simulate single-cell MTX data for DEG analysis}
\usage{
make.sc.deg(
  file.header,
  nind = 40,
  ngene = 1000,
  ncausal = 5,
  ncovar.conf = 3,
  ncovar.batch = 0,
  ncell.ind = 10,
  pve.1 = 0.3,
  pve.c = 0.5,
  pve.a = 0.5,
  rho.a = 2,
  rho.b = 2,
  rseed = 13,
  exposure.type = c("binary", "continuous")
)
}
\arguments{
\item{file.header}{file set header}

\item{nind}{num of individuals}

\item{ngene}{num of genes/features}

\item{ncausal}{num of causal genes}

\item{ncovar.conf}{num of confounding covariates}

\item{ncovar.batch}{num of confounding batch variables}

\item{ncell.ind}{num of cells per individual}

\item{pve.1}{variance of treatment/disease effect}

\item{pve.c}{variance of confounding effect}

\item{pve.a}{variance of confounders to the assignment}

\item{rho.a}{rho ~ gamma(a, b)}

\item{rho.b}{rho ~ gamma(a, b)}

\item{rseed}{random seed}

\item{exposure.type}{"binary" or "continuous"}
}
\value{
simulation results
}
\description{
Simulate single-cell MTX data for DEG analysis
}
\details{
The simulation result list will have two lists:

\code{data}:
\itemize{
\item a matrix market data file \code{data$mtx}
\item a file with row names \code{data$row}
\item a file with column names \code{data$col}
\item an indexing file for the columns \code{data$idx}
\item a mapping file between column and individual names "indv"
}

\code{indv}:
\itemize{
\item \code{obs.mu} observed (noisy) gene x individual matrix
\item \code{clean.mu} clean gene x individual matrix
\item \code{X} confounder x individual matrix
\item \code{W} individual-level treatment assignment
\item \code{rho} sequencing depth
\item \code{causal} causal genes
}
}
