% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{make.sc.eqtl}
\alias{make.sc.eqtl}
\title{Simulate individual-level eQTL data}
\usage{
make.sc.eqtl(
  file.header,
  X,
  h2,
  n.causal.snps = 1,
  n.causal.genes = 5,
  pve.y.by.u0 = 0.3,
  n.u0 = 3,
  pve.u1.by.x = 0.8,
  pve.y.by.u1 = 0.3,
  n.u1 = 3,
  pve.interaction = 0.5,
  n.interaction = 0,
  n.genes = 50,
  rho.a = 2,
  rho.b = 2,
  ncell.ind = 10,
  ind.prob = NULL
)
}
\arguments{
\item{X}{genotype matrix (individual x SNPs)}

\item{h2}{heritability (proportion of variance of Y explained by genetic X)}

\item{n.causal.snps}{X variables directly affecting on Y}

\item{n.causal.genes}{Y variables directly regulated by X}

\item{pve.y.by.u0}{proportion of variance of Y explained by U0}

\item{n.u0}{number of covariates on Y}

\item{pve.u1.by.x}{proportion of variance of U1 explained by X}

\item{pve.y.by.u1}{proportion of variance of Y explained by U1}

\item{n.u1}{number of covariates on Y}

\item{pve.interaction}{proportion of variance of Y explained by interaction}

\item{n.interaction}{number of genes interacting with the causal genes}

\item{n.genes}{total number of genes (Y variables)}

\item{rho.a}{rho ~ gamma(a, b)}

\item{rho.b}{rho ~ gamma(a, b)}

\item{ncell.ind}{number of cells per individual}
}
\value{
simulation results
}
\description{
The goal is to have unbiased estimates of Y ~ X
}
\details{
However, possible confounding effect models lurking:

Y1 ~ X + U
Y0 ~ Y1 + U

The simulation result list will have two lists:

\code{data}:
\itemize{
\item \code{data$mtx}: a matrix market data file
\item \code{data$row}: a file with row names
\item \code{data$col}: a file with column names
\item \code{data$idx}: an indexing file for the columns
\item \code{data$indv}: a mapping file between column and individual names
}

\code{indv}:
\itemize{
\item \code{indv$y}: observed (noisy) individual x gene matrix
\item \code{indv$x}: observed individual x variants genotype matrix
\item \code{indv$causal.snps}: causal variants (X variables)
\item \code{indv$causal.genes}: causal genes (Y variables)
}
}
