% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{pine.remove.duplicated}
\alias{pine.remove.duplicated}
\title{Check duplicated pairs of individuals in the matching results
and take the nearest ones}
\usage{
pine.remove.duplicated(input)
}
\arguments{
\item{input}{the whole result of \code{make.pine}}
}
\value{
filtered \code{input} results
}
\description{
Check duplicated pairs of individuals in the matching results
and take the nearest ones
}
