% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_build_mmutil_index}
\alias{rcpp_build_mmutil_index}
\title{Create an index file for a given MTX}
\usage{
rcpp_build_mmutil_index(mtx_file, index_file)
}
\arguments{
\item{mtx_file}{data file}

\item{index_file}{index file}
}
\value{
EXIT_SUCCESS or EXIT_FAILURE
}
\description{
Create an index file for a given MTX
}
