% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_build_sparse_mat}
\alias{rcpp_build_sparse_mat}
\title{Build sparse matrix from triplets}
\usage{
rcpp_build_sparse_mat(A_ij_list, Nrow, Ncol, symmetrize = TRUE)
}
\arguments{
\item{A_ij_list}{list(src.index, tgt.index, \link{weights})}

\item{Nrow}{number of rows}

\item{Ncol}{number of columns}

\item{symmetrize}{symmetrize A matrix}
}
\value{
a sparse matrix A
}
\description{
Build sparse matrix from triplets
}
