% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_check_index_tab}
\alias{rcpp_check_index_tab}
\title{Check if the index tab is valid}
\usage{
rcpp_check_index_tab(mtx_file, index_tab)
}
\arguments{
\item{mtx_file}{data file}

\item{index_tab}{index tab (a vector of memory locations)}
}
\value{
EXIT_SUCCESS or EXIT_FAILURE
}
\description{
Check if the index tab is valid
}
