% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_compute_scores}
\alias{rcpp_compute_scores}
\title{Collect row-wise and column-wise statistics}
\usage{
rcpp_compute_scores(mtx_file)
}
\arguments{
\item{mtx_file}{data file}
}
\value{
a list of stat vectors
}
\description{
Collect row-wise and column-wise statistics
}
\examples{
rr <- rgamma(10, 1, 1) # ten cells
mm <- matrix(rgamma(10 * 3, 1, 1), 10, 3)
dat <- mmutilR::rcpp_simulate_poisson_data(mm, rr, "sim_test")
scr <- mmutilR::rcpp_compute_scores(dat$mtx)
A <- as.matrix(Matrix::readMM(dat$mtx))
colMeans(A)
scr$col.mean
rowMeans(A)
scr$row.mean

}
