% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_copy_selected_columns}
\alias{rcpp_copy_selected_columns}
\title{Take a subset of columns and create a new MTX file-set}
\usage{
rcpp_copy_selected_columns(mtx_file, row_file, col_file, r_selected, output)
}
\arguments{
\item{mtx_file}{data file}

\item{row_file}{row file}

\item{col_file}{column file}

\item{selected}{selected column names}
}
\description{
Take a subset of columns and create a new MTX file-set
}
\examples{

options(stringsAsFactors=FALSE)
rr <- rgamma(20, 1, 1)
mm <- matrix(rgamma(10 * 2, 1, 1), 10, 2)
src.hdr <- "test_org"
src.files <- mmutilR::rcpp_simulate_poisson_data(mm, rr, src.hdr)
Y <- Matrix::readMM(src.files$mtx)
colnames(Y) <- read.table(src.files$col)$V1
print(Y)
sub.cols <- sort(read.table(src.files$col)$V1[sample(20,3)])
print(sub.cols)
tgt.hdr <- "test_sub"
tgt.files <- mmutilR::rcpp_copy_selected_columns(src.files$mtx,
                                     src.files$row,
                                     src.files$col,
                                     sub.cols, tgt.hdr)
Y <- Matrix::readMM(tgt.files$mtx)
colnames(Y) <- read.table(tgt.files$col)$V1
print(Y)
unlink(list.files(pattern = src.hdr))
unlink(list.files(pattern = tgt.hdr))

}
