% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_copy_selected_rows}
\alias{rcpp_copy_selected_rows}
\title{Take a subset of rows and create a new MTX file-set}
\usage{
rcpp_copy_selected_rows(mtx_file, row_file, col_file, r_selected, output)
}
\arguments{
\item{mtx_file}{data file}

\item{row_file}{row file}

\item{col_file}{column file}

\item{output}{output}

\item{selected}{selected row names}
}
\value{
a list of file names: {output}.{mtx,rows,cols}.gz
}
\description{
For the new mtx file, empty columns with only zero
  elements will be removed.
}
\examples{

options(stringsAsFactors=FALSE)
rr <- rgamma(20, 1, 1)
mm <- matrix(rgamma(10 * 2, 1, 1), 10, 2)
src.hdr <- "test_org"
src.files <- mmutilR::rcpp_simulate_poisson_data(mm, rr, src.hdr)
Y <- Matrix::readMM(src.files$mtx)
rownames(Y) <- read.table(src.files$row)$V1
print(Y)
sub.rows <- sort(read.table(src.files$row)$V1[sample(10,3)])
print(sub.rows)
tgt.hdr <- "test_sub"
tgt.files <- mmutilR::rcpp_copy_selected_rows(src.files$mtx,
                                              src.files$row,
                                              src.files$col,
                                              sub.rows,
                                              tgt.hdr)
Y <- Matrix::readMM(tgt.files$mtx)
colnames(Y) <- read.table(tgt.files$col)$V1
rownames(Y) <- read.table(tgt.files$row)$V1
print(Y)
unlink(list.files(pattern = src.hdr))
unlink(list.files(pattern = tgt.hdr))

}
