% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_aggregate_pairwise}
\alias{rcpp_mmutil_aggregate_pairwise}
\title{Create pseudo-bulk data for pairwise comparisons}
\usage{
rcpp_mmutil_aggregate_pairwise(
  mtx_file,
  row_file,
  col_file,
  r_indv,
  r_V,
  r_cols = NULL,
  r_annot = NULL,
  r_annot_mat = NULL,
  r_lab_name = NULL,
  a0 = 1,
  b0 = 1,
  eps = 1e-08,
  knn_cell = 10L,
  knn_indv = 1L,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  NUM_THREADS = 1L,
  IMPUTE_BY_KNN = TRUE,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{data file}

\item{row_file}{row file}

\item{col_file}{column file}

\item{r_indv}{membership for the cells (\code{r_cols})}

\item{r_V}{SVD factors}

\item{r_cols}{cell (col) names (if we want to take a subset)}

\item{r_annot}{label annotation for the (\code{r_cols})}

\item{r_annot_mat}{label annotation matrix (cell x type) (default: NULL)}

\item{r_lab_name}{label names (default: everything in \code{r_annot})}

\item{a0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{b0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{eps}{small number (default: 1e-8)}

\item{knn_cell}{k-NN matching between cells}

\item{knn_indv}{k-NN matching between individuals}

\item{KNN_BILINK}{num. of bidirectional links (default: 10)}

\item{KNN_NNLIST}{num. of nearest neighbor lists (default: 10)}

\item{NUM_THREADS}{number of threads for multi-core processing}

\item{IMPUTE_BY_KNN}{imputation by kNN alone (default: TRUE)}
}
\value{
a list of inference results
}
\description{
Create pseudo-bulk data for pairwise comparisons
}
