% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_bbknn_pca}
\alias{rcpp_mmutil_bbknn_pca}
\title{BBKNN(Batch-balancing kNN)-adjusted PCA}
\usage{
rcpp_mmutil_bbknn_pca(
  mtx_file,
  r_batches,
  knn,
  RANK,
  TAKE_LN = TRUE,
  TAU = 1,
  COL_NORM = 10000,
  EM_ITER = 10L,
  EM_TOL = 1e-04,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  LU_ITER = 5L,
  row_weight_file = ""
)
}
\arguments{
\item{mtx_file}{data file (feature x n)}

\item{r_batches}{batch names (n x 1)}

\item{knn}{kNN parameter k}

\item{RANK}{SVD rank}

\item{TAKE_LN}{take log(1 + x) trans or not}

\item{TAU}{regularization parameter (default = 1)}

\item{COL_NORM}{column normalization}

\item{EM_ITER}{EM iteration for factorization (default: 10)}

\item{EM_TOL}{EM convergence (default: 1e-4)}

\item{KNN_BILINK}{# of bidirectional links (default: 10)}

\item{KNN_NNLIST}{# nearest neighbor lists (default: 10)}

\item{LU_ITER}{LU iteration}

\item{row_weight_file}{row-wise weight file}
}
\value{
a list of (1) U (2) D (3) V
}
\description{
BBKNN(Batch-balancing kNN)-adjusted PCA
}
\examples{
## Generate some data
rr <- rgamma(1000, 1, 1) # 1000 cells
mm <- matrix(rgamma(100 * 3, 1, 1), 100, 3)
dat <- mmutilR::rcpp_simulate_poisson_data(mm, rr, "sim_test")
.ind <- read.table(dat$indv)
.col <- unlist(read.table(dat$col))
.ind <- .ind[match(.col, .ind$V1), ]
.bbknn <- mmutilR::rcpp_mmutil_bbknn_pca(dat$mtx,
                  .ind$V2, 10, 3, TAKE_LN = FALSE)
plot(.bbknn$V[, 1], .bbknn$V[, 2], col = .ind$V2,
     xlab = "PC1", ylab = "PC2")
plot(.bbknn$factors.adjusted[, 1], .bbknn$factors.adjusted[, 2],
     col = .ind$V2,
     xlab = "PC1 (BBKNN)", ylab = "PC2 (BBKNN)")
## clean up temp directory
unlink(list.files(pattern = "sim_test"))

}
