% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_colnames}
\alias{rcpp_mmutil_colnames}
\title{Just read each col name per line}
\usage{
rcpp_mmutil_colnames(col_file, MAX_COL_WORD = 100L, COL_WORD_SEP = "@")
}
\arguments{
\item{col_file}{data file}
}
\value{
colnames
}
\description{
Just read each col name per line
}
