% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_deconvolve_svd}
\alias{rcpp_mmutil_deconvolve_svd}
\title{Cell type deconvolution of bulk data based on single-cell data}
\usage{
rcpp_mmutil_deconvolve_svd(
  mtx_file = "",
  row_file = "",
  col_file = "",
  r_U = NULL,
  r_D = NULL,
  r_V = NULL,
  r_rows = NULL,
  r_cols = NULL,
  TAKE_LN = FALSE,
  VERBOSE = FALSE
)
}
\description{
Salamander (semi-supervised annotation of latent states by marker
gene-derived regression model)
}
