% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_merge_file_sets}
\alias{rcpp_mmutil_merge_file_sets}
\title{Merge multiple 10x mtx file sets into one set}
\usage{
rcpp_mmutil_merge_file_sets(
  r_headers = NULL,
  r_batches = NULL,
  r_mtx_files = NULL,
  r_row_files = NULL,
  r_col_files = NULL,
  r_fixed_rows = NULL,
  output = "output",
  nnz_cutoff = 1,
  delim = "_"
)
}
\arguments{
\item{r_headers}{file set headers}

\item{r_batches}{unique batch names for each header}

\item{r_mtx_files}{A list of mtx files}

\item{r_row_files}{A list of row files}

\item{r_col_files}{A list of col files}

\item{r_fixed_rows}{A list of rows/features}

\item{output}{output file header}

\item{nnz_cutoff}{number of non-zero cutoff for columns}

\item{delim}{delimiter in the column name}
}
\value{
a list of file names: {output}.{mtx,rows,cols}.gz
}
\description{
Merge multiple 10x mtx file sets into one set
}
\examples{

options(stringsAsFactors=FALSE)
rr <- rgamma(10, 1, 1) # ten cells
mm <- matrix(rgamma(10 * 3, 1, 1), 10, 3)
t1 <- mmutilR::rcpp_mmutil_simulate_poisson(mm, rr, "test1")
t2 <- mmutilR::rcpp_mmutil_simulate_poisson(mm, rr, "test2")
bats <- hdrs <- c("test1","test2")
t3 <- mmutilR::rcpp_mmutil_merge_file_sets(
                      hdrs, bats, "test3", 0)
A1 <- Matrix::readMM(t1$mtx);
rownames(A1) <- unlist(read.table(gzfile(t1$row)))
A2 <- Matrix::readMM(t2$mtx)
rownames(A2) <- unlist(read.table(gzfile(t2$row)))
A3 <- Matrix::readMM(t3$mtx)
rownames(A3) <- unlist(read.table(gzfile(t3$row)))
print(cbind(A1, A2))
print(A3)
unlink(list.files(pattern = "test1"))
unlink(list.files(pattern = "test2"))
unlink(list.files(pattern = "test3"))

}
