% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_network_edge_cluster}
\alias{rcpp_mmutil_network_edge_cluster}
\title{Clustering columns of the network mtx file (feature incidence matrix)}
\usage{
rcpp_mmutil_network_edge_cluster(
  mtx_file,
  row_file,
  col_file,
  output,
  num_clust = 10L,
  num_gibbs = 100L,
  num_burnin = 100L,
  nnz_cutoff = 1L,
  A0 = 1,
  B0 = 1,
  Dir0 = 1,
  verbose = TRUE
)
}
\arguments{
\item{mtx_file}{data file (feature x edge)}

\item{row_file}{row file (feature x 1)}

\item{col_file}{col file (sample x 1)}

\item{output}{a file header for result/temporary files}

\item{nnz_cutoff}{Only consider edge with NNZ >= nnz_cutoff (default: 1)}
}
\description{
Clustering columns of the network mtx file (feature incidence matrix)
}
