% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_network_topic_data}
\alias{rcpp_mmutil_network_topic_data}
\title{Construct a kNN cell-cell interaction network and identify gene topics}
\usage{
rcpp_mmutil_network_topic_data(
  mtx_file,
  row_file,
  col_file,
  latent_factor,
  knn,
  output,
  write_sample_network = FALSE,
  output_sample_incidence = NULL,
  output_sample_adjacency = NULL,
  r_batches = NULL,
  CUTOFF = 0,
  WEIGHTED = TRUE,
  MAXW = 1,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  NUM_THREADS = 1L,
  MAX_ROW_WORD = 2L,
  ROW_WORD_SEP = "_",
  MAX_COL_WORD = 100L,
  COL_WORD_SEP = "@"
)
}
\arguments{
\item{mtx_file}{data file (feature x n)}

\item{row_file}{row file (feature x 1)}

\item{col_file}{row file (n x 1)}

\item{latent_factor}{(n x K)}

\item{knn}{kNN parameter}

\item{output}{file header for resulting files}

\item{write_sample_network}{Do we want sample inc/adj? (default: FALSE)}

\item{output_sample_incidence}{file header for sample adj (default: NULL)}

\item{r_batches}{batch names (n x 1, default: NULL)}

\item{CUTOFF}{expression present/absent call cutoff (default: 0)}

\item{WEIGHTED}{include edge weights in feature calculation (default: TRUE)}

\item{MAXW}{maximum edge weight (default: 1)}

\item{KNN_BILINK}{num. of bidirectional links (default: 10)}

\item{KNN_NNLIST}{num. of nearest neighbor lists (default: 10)}

\item{NUM_THREADS}{number of threads for multi-core processing}

\item{MAX_ROW_WORD}{maximum words per line in \code{row_file}}

\item{ROW_WORD_SEP}{word separation character to replace white space}

\item{MAX_COL_WORD}{maximum words per line in \code{col_file}}

\item{COL_WORD_SEP}{word separation character to replace white space}
}
\value{
feature.incidence, sample.incidence, sample.adjacency
}
\description{
Construct a kNN cell-cell interaction network and identify gene topics
}
