% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_pca}
\alias{rcpp_mmutil_pca}
\title{Approximate PCA}
\usage{
rcpp_mmutil_pca(
  mtx_file,
  RANK,
  TAKE_LN = TRUE,
  TAU = 1,
  COL_NORM = 10000,
  EM_ITER = 0L,
  EM_TOL = 1e-04,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  LU_ITER = 5L,
  row_weight_file = "",
  NUM_THREADS = 1L,
  BLOCK_SIZE = 10000L
)
}
\arguments{
\item{mtx_file}{data file (feature x n)}

\item{RANK}{SVD rank}

\item{TAKE_LN}{take log(1 + x) trans or not}

\item{TAU}{regularization parameter (default = 1)}

\item{COL_NORM}{column normalization}

\item{EM_ITER}{EM iteration for factorization (default: 10)}

\item{EM_TOL}{EM convergence (default: 1e-4)}

\item{LU_ITER}{LU iteration}

\item{row_weight_file}{row-wise weight file}

\item{NUM_THREADS}{number of threads for multi-core processing}

\item{BLOCK_SIZE}{number of columns per block}
}
\value{
a list of (1) U (2) D (3) V
}
\description{
Approximate PCA
}
\examples{
## Generate some data
set.seed(1)
rr <- rgamma(1000, 1, 1) # 1000 cells
mm <- matrix(rgamma(100 * 3, 1, 1), 100, 3)
.dat <- mmutilR::rcpp_mmutil_simulate_poisson(mm, rr, "sim_test")
.pc <- mmutilR::rcpp_mmutil_pca(.dat$mtx, 3, TAKE_LN = FALSE)
.ind <- read.table(.dat$indv)
.col <- unlist(read.table(.dat$col))
.ind <- .ind[match(.col, .ind$V1), ]
plot(.pc$V[, 1], .pc$V[, 2], col = .ind$V2,
     xlab = "PC1", ylab = "PC2")
plot(.pc$V[, 2], .pc$V[, 3], col = .ind$V2,
     xlab = "PC2", ylab = "PC3")
## clean up temp directory
unlink(list.files(pattern = "sim_test"))

}
