% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_read_columns_sparse}
\alias{rcpp_mmutil_read_columns_sparse}
\title{Read a subset of columns from the data matrix}
\usage{
rcpp_mmutil_read_columns_sparse(
  mtx_file,
  memory_location,
  r_column_index,
  verbose = FALSE,
  NUM_THREADS = 1L,
  MIN_SIZE = 10L
)
}
\arguments{
\item{mtx_file}{data file}

\item{memory_location}{column -> memory location}

\item{r_column_index}{column indexes to retrieve (1-based)}
}
\value{
lists of rows, columns, values
}
\description{
Read a subset of columns from the data matrix
}
