% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_rownames}
\alias{rcpp_mmutil_rownames}
\title{Just read each row name per line}
\usage{
rcpp_mmutil_rownames(row_file, MAX_ROW_WORD = 2L, ROW_WORD_SEP = "_")
}
\arguments{
\item{row_file}{data file}
}
\value{
rownames
}
\description{
Just read each row name per line
}
