% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_simulate_poisson_mixture}
\alias{rcpp_mmutil_simulate_poisson_mixture}
\title{Simulate sparse counting data with a mixture of Poisson parameters}
\usage{
rcpp_mmutil_simulate_poisson_mixture(
  r_mu_list,
  Ncell,
  output,
  dir_alpha = 1,
  gam_alpha = 2,
  gam_beta = 2,
  rseed = 42L
)
}
\arguments{
\item{r_mu_list}{a list of gene x individual matrices}

\item{Ncell}{the total number of cells (may not make it if too sparse)}

\item{output}{a file header string for output files}

\item{dir_alpha}{a parameter for Dirichlet(alpha * \link{1, ..., 1})}

\item{gam_alpha}{a parameter for Gamma(alpha, beta)}

\item{gam_beta}{a parameter for Gamma(alpha, beta)}

\item{rseed}{random seed}
}
\description{
Simulate sparse counting data with a mixture of Poisson parameters
}
