% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_read_columns}
\alias{rcpp_read_columns}
\title{Read a subset of columns from the data matrix}
\usage{
rcpp_read_columns(mtx_file, memory_location, r_column_index)
}
\arguments{
\item{mtx_file}{data file}

\item{memory_location}{column -> memory location}

\item{r_column_index}{column indexes to retrieve (1-based)}
}
\value{
a dense sub-matrix
}
\description{
Read a subset of columns from the data matrix
}
\examples{

rr <- rgamma(100, 1, 1) # one hundred cells
mm <- matrix(rgamma(10 * 3, 1, 1), 10, 3)
data.hdr <- "test_sim"
.files <- mmutilR::rcpp_simulate_poisson_data(mm, rr, data.hdr)
data.file <- .files$mtx
idx.file <- .files$idx
mtx.idx <- mmutilR::rcpp_read_mmutil_index(idx.file)
Y <- as.matrix(Matrix::readMM(data.file))
col.pos <- c(1,13,77) # 1-based
yy <- mmutilR::rcpp_read_columns(data.file, mtx.idx, col.pos)
all(Y[, col.pos, drop = FALSE] == yy)
print(head(Y[, col.pos, drop = FALSE]))
print(head(yy))
unlink(list.files(pattern = data.hdr))

}
