% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{read.mtx.sparse}
\alias{read.mtx.sparse}
\title{A wrapper function to read a sparse submatrix}
\usage{
read.mtx.sparse(
  mtx.file,
  row.file = NULL,
  col.file = NULL,
  sub.cols = NULL,
  memory.idx = NULL,
  memory.idx.file = paste0(mtx.file, ".index"),
  verbose = FALSE,
  NUM_THREADS = 1,
  max.row.word = 2,
  row.word.sep = "_",
  max.col.word = 100,
  col.word.sep = "@"
)
}
\arguments{
\item{mtx.file}{matrix market file}

\item{row.file}{row file (default: NULL)}

\item{col.file}{column file (default: NULL)}

\item{sub.cols}{column index (default: NULL, all)}

\item{memory.idx}{memory locations}

\item{memory.idx.file}{memory location file}

\item{verbose}{verbosity}

\item{max.row.word}{maximum number of words per each row (default: 2)}

\item{row.word.sep}{row word separator character (default: "_")}

\item{max.col.word}{maximum number of words per each col (default: 100)}

\item{col.word.sep}{column word separator character (default: "@")}
}
\value{
a sparse matrix
}
\description{
A wrapper function to read a sparse submatrix
}
