% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulate.data}
\alias{simulate.data}
\title{Simulate single-cell MTX data}
\usage{
\method{simulate}{data}(file.header, ...)
}
\arguments{
\item{file.header}{file set header}

\item{nind}{# individuals}

\item{ngene}{# genes/features}

\item{ncausal}{# causal genes}

\item{ncovar.conf}{# confounding covariates}

\item{ncovar.batch}{# confounding batch variables}

\item{ncell.ind}{# cells per individual}

\item{pve.1}{variance of treatment/disease effect}

\item{pve.c}{variance of confounding effect}

\item{pve.a}{variance of confounders to the assignment}

\item{rho.a}{rho ~ gamma(a, b)}

\item{rho.b}{rho ~ gamma(a, b)}

\item{rseed}{random seed}

\item{exposure.type}{"binary" or "continuous"}
}
\value{

}
\description{
Simulate single-cell MTX data
}
