% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulate.eqtl.data}
\alias{simulate.eqtl.data}
\title{Simulate individual-level eQTL data}
\usage{
\method{simulate}{eqtl.data}(
  file.header,
  rho.a = 2,
  rho.b = 2,
  ncell.ind = 10,
  ind.prob = NULL,
  ...
)
}
\arguments{
\item{rho.a}{rho ~ gamma(a, b)}

\item{rho.b}{rho ~ gamma(a, b)}

\item{ncell.ind}{number of cells per individual}

\item{X}{genotype matrix (individual x SNPs)}

\item{h2}{heritability (proportion of variance of Y explained by genetic X)}

\item{pve.u1.by.x}{(proportion of variance of U1 explained by X)}

\item{pve.y.by.u1}{(proportion of variance of Y explained by U1)}

\item{pve.y.by.u0}{(proportion of variance of Y explained by U0)}

\item{n.causal.snps}{#(X directly affecting on Y)}

\item{n.causal.genes}{#(Y regulated by X)}

\item{n.u1}{#(U1 variables)}

\item{n.u0}{#(U0 variables)}

\item{n.genes}{total number of genes (Y variables)}
}
\value{
simulation results
}
\description{
Possible confounding effect model:
}
\details{
X -> U1 -> Y : harmful if U1 was adjusted (epigenetics, gene regulation, trans-effects)
U0 -> Y : okay to adjust to boost the power (demographic variables, environments)

Another rare possibility (we don't consider here):

X -> Y -> Z : harmful if Z was adjusted (downstream phenotypes mediated by genes)

The simulation result list will have two lists:

\code{data}:
\itemize{
\item \code{data$mtx}: a matrix market data file
\item \code{data$row}: a file with row names
\item \code{data$col}: a file with column names
\item \code{data$idx}: an indexing file for the columns
\item \code{data$indv}: a mapping file between column and individual names
}

\code{indv}:
\itemize{
\item \code{indv$obs.mu}: observed (noisy) gene x individual matrix
\item \code{indv$clean.mu}: clean gene x individual matrix
\item \code{indv$X}: confounder x individual matrix
\item \code{indv$W}: individual-level treatment assignment,
\item \code{indv$rho}: sequencing depth
\item \code{indv$causal}: causal genes
}
}
