% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulate_gamma_glm}
\alias{simulate_gamma_glm}
\title{Simulate individual-level effects by GLM}
\usage{
simulate_gamma_glm(
  nind = 40,
  ngene = 1000,
  ncausal = 5,
  ncovar.conf = 1,
  ncovar.batch = 3,
  ncell.ind = 10,
  pve.1 = 0.3,
  pve.c = 0.5,
  pve.a = 0.5,
  rho.a = 2,
  rho.b = 2,
  rseed = 13
)
}
\arguments{
\item{nind}{number of individuals/samples}

\item{ngene}{# genes/features}

\item{ncausal}{number of causal genes}

\item{ncovar.conf}{number of covar shared}

\item{ncovar.batch}{number of covar on mu, batch effect}

\item{ncell.ind}{# cells per individual}

\item{pve.1}{variance of treatment/disease effect}

\item{pve.c}{variance of confounding effect}

\item{pve.a}{variance of confounders to the assignment}

\item{rho.a}{rho ~ gamma(a, b)}

\item{rho.b}{rho ~ gamma(a, b)}

\item{rseed}{random seed

simple concat
random standard Normal}

\item{n2}{sample model parameters}

\item{ngenes}{number of genes/features}
}
\description{
Simulate individual-level effects by GLM
}
