% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulate_indv_eqtl}
\alias{simulate_indv_eqtl}
\title{Simulate individual-level eQTL data}
\usage{
simulate_indv_eqtl(
  X,
  h2,
  n.causal.snps,
  n.causal.genes,
  pve.y.by.u0,
  n.u0,
  pve.u1.by.x,
  pve.y.by.u1,
  n.u1,
  pve.interaction,
  n.interaction,
  n.genes
)
}
\arguments{
\item{X}{genotype matrix (individual x SNPs)}

\item{h2}{heritability (proportion of variance of Y explained by genetic X)}

\item{n.causal.snps}{X variables directly affecting on Y}

\item{n.causal.genes}{Y variables directly regulated by X}

\item{pve.y.by.u0}{proportion of variance of Y explained by U0}

\item{n.u0}{number of covariates on Y}

\item{pve.u1.by.x}{proportion of variance of U1 explained by X}

\item{pve.y.by.u1}{proportion of variance of Y explained by U1}

\item{n.u1}{number of covariates on Y}

\item{pve.interaction}{proportion of variance of Y explained by interaction}

\item{n.interaction}{number of genes interacting with the causal genes}

\item{n.genes}{total number of genes (Y variables)}
}
\value{
simulation results
}
\description{
Simulate individual-level eQTL data
}
\details{

}
