% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{vcat.sparse}
\alias{vcat.sparse}
\title{A wrapper function that concatenates two files vertically}
\usage{
vcat.sparse(top.hdr, bottom.hdr, out.hdr)
}
\arguments{
\item{top.hdr}{a file header for the top files}

\item{bottom.hdr}{a file header for the bottom files}

\item{out.hdr}{output file header}
}
\value{
a list of the resulting file names
}
\description{
A wrapper function that concatenates two files vertically
}
