% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{write.sparse}
\alias{write.sparse}
\title{Write matrix market file set}
\usage{
write.sparse(out.mtx, out.rows, out.cols, output)
}
\arguments{
\item{out.mtx}{a sparse matrix}

\item{out.rows}{a vector of rows}

\item{out.cols}{a vector of columns}

\item{output}{output file set header}
}
\description{
Write matrix market file set
}
