% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{make.pine}
\alias{make.pine}
\title{Counterfactual confounder adjustment by
by Pairwise INdividual Effect matching}
\usage{
make.pine(
  mtx.data,
  celltype,
  cell2indv,
  V = NULL,
  knn.cell = 50,
  knn.indv = 1,
  celltype.mat = NULL,
  .rank = 10,
  .take.ln = TRUE,
  .pca.reg = 1,
  .col.norm = 10000,
  .em.iter = 0,
  .em.tol = 1e-04,
  num.threads = 1,
  impute.by.knn = FALSE,
  remove.dup = TRUE,
  ...
)
}
\arguments{
\item{mtx.data}{fileset.header ($mtx, $row, $col, $idx)}

\item{celltype}{celltype/cluster assignments (cells x 2 mapping, cells x 1, or just a single string)}

\item{cell2indv}{cell-level individual assignments (cells x 2), cell -> indv}

\item{celltype.mat}{celltype/cluster assignment matrix (cells x cell types)}

\item{.rank}{SVD rank for spectral matching}

\item{.take.ln}{take log(1 + x) trans or not}

\item{.pca.reg}{regularization parameter (default = 1)}

\item{.col.norm}{column normalization for SVD}

\item{.em.iter}{EM iteration for factorization (default: 0)}

\item{.em.tol}{EM convergence (default: 1e-4)}

\item{num.threads}{number of threads for multi-core processing}

\item{impute.by.knn}{imputation by kNN weighting (default: FALSE)}

\item{remove.dup}{remove duplicated pairs (default: TRUE)}

\item{eps}{small number (default: 1e-8)}

\item{knn}{number of neighbours \code{k} in kNN for matching}

\item{a0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{b0}{hyperparameter for gamma(a0, b0) (default: 1)}
}
\value{
a list of sufficient statistics matrices
}
\description{
Counterfactual confounder adjustment by
by Pairwise INdividual Effect matching
}
\examples{

sim.data <- make.sc.deg.data("temp",
                             nind = 40,
                             ngenes = 100,
                             ngenes.covar = 50,
                             ncausal = 3,
                             ncovar.conf = 3,
                             ncovar.batch = 0,
                             ncell.ind = 20,
                             pve.1 = .3,
                             pve.c = .5,
                             pve.a = .5,
                             rseed = 13,
                             rho.a = 2,
                             rho.b = 2,
                             exposure.type = "continuous",
                             smudge = 5)

cell2indv <- read.table(sim.data$data$indv,
                        header=FALSE,
                        col.names=c("cell","indv"))

.pine <- make.pine(sim.data$data,
                   "bulk",
                   cell2indv,
                   knn.cell = 10,
                   knn.indv = 10)

Yd <- t(.pine$ln.delta)
uu <- t(.pine$covar.matched)
Yd <- t(lm(Yd ~ uu)$residuals)
Yd0 <- .pine$ln.delta

W <- sim.data$indv$assignment

.src <- .pine$knn$obs.index
.tgt <- .pine$knn$matched.index

w.delta <- W[.src] - W[.tgt]

causal <- sim.data$causal

ngenes <- nrow(Yd)
p.values <- apply(Yd, 1, function(y) { cor.test(w.delta, y)$p.value })
p.values.0 <- apply(Yd0, 1, function(y) { cor.test(w.delta, y)$p.value })

par(mfrow=c(2,1))
plot(-log10(p.values), col="gray40", main = "adjusted")
points(causal, -log10(p.values[causal]), col = 2, pch=19)

plot(-log10(p.values.0), col="gray40", main = "unadjusted")
points(causal, -log10(p.values.0[causal]), col = 2, pch=19)

unlink(list.files(pattern = "temp"))

}
