% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{make.pine}
\alias{make.pine}
\title{Counterfactual confounder adjustment by
by Pairwise INdividual Effect matching}
\usage{
make.pine(
  mtx.data,
  celltype,
  cell2indv,
  knn.cell = 50,
  knn.indv = 1,
  celltype.mat = NULL,
  .rank = 10,
  .take.ln = TRUE,
  .pca.reg = 1,
  .col.norm = 10000,
  .em.iter = 0,
  .em.tol = 1e-04,
  num.threads = 1,
  remove.dup = TRUE,
  ...
)
}
\arguments{
\item{mtx.data}{fileset.header ($mtx, $row, $col, $idx)}

\item{celltype}{celltype/cluster assignments (1 x cells or just a single string)}

\item{cell2indv}{cell-level individual assignments (cells x 2), cell -> indv}

\item{celltype.mat}{celltype/cluster assignment matrix (cells x cell types)}

\item{.rank}{SVD rank for spectral matching}

\item{.take.ln}{take log(1 + x) trans or not}

\item{.pca.reg}{regularization parameter (default = 1)}

\item{.col.norm}{column normalization for SVD}

\item{.em.iter}{EM iteration for factorization (default: 0)}

\item{.em.tol}{EM convergence (default: 1e-4)}

\item{num.threads}{number of threads for multi-core processing}

\item{eps}{small number (default: 1e-8)}

\item{knn}{number of neighbours \code{k} in kNN for matching}

\item{a0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{b0}{hyperparameter for gamma(a0, b0) (default: 1)}
}
\value{
a list of sufficient statistics matrices
}
\description{
Counterfactual confounder adjustment by
by Pairwise INdividual Effect matching
}
\examples{

sim.data <- make.sc.deg("temp",
                              nind = 20,
                              ngene = 100,
                              ncausal = 3,
                              ncovar.conf = 3,
                              ncovar.batch = 0,
                              ncell.ind = 10,
                              pve.1 = .3,
                              pve.c = .5,
                              pve.a = .5,
                              rseed = 13,
                              exposure.type = "continuous")

mtx.data <- sim.data$data

cell2indv <- read.table(sim.data$data$indv,
                        header=FALSE,
                        col.names=c("cell","indv"))

nind <- length(sim.data$indv$W)

.pine <- make.pine(mtx.data,
                   "bulk",
                   cell2indv,
                   knn.cell = 50,
                   knn.indv = 1)

.names <- lapply(colnames(.pine$delta), strsplit, split="[_]")
.names <- lapply(.names, function(x) unlist(x))
.pairs <- data.frame(do.call(rbind, .names), stringsAsFactors=FALSE)
colnames(.pairs) <- c("src","tgt","ct")

W <- sim.data$indv$W
w.src <- W[as.integer(.pairs$src)]
w.tgt <- W[as.integer(.pairs$tgt)]
w.delta <- w.src - w.tgt

.pine$delta[.pine$delta < 0] <- NA # numerical errors

ncausal <- length(sim.data$indv$causal)
ngene <- nrow(.pine$delta)

par(mfrow=c(2, ncausal))
for(k in sim.data$indv$causal){
    plot(w.delta, .pine$delta[k, ], xlab = "dW", ylab = "dY")
}

for(k in sample(setdiff(1:ngene, sim.data$indv$causal), ncausal)){
    plot(w.delta, .pine$delta[k, ], xlab = "dW", ylab = "dY")
}

unlink(list.files(pattern = "temp"))

}
