% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_aggregate}
\alias{rcpp_mmutil_aggregate}
\title{Create pseudo-bulk data by aggregating columns}
\usage{
rcpp_mmutil_aggregate(
  mtx_file,
  row_file,
  col_file,
  r_cols = NULL,
  r_indv = NULL,
  r_annot = NULL,
  r_annot_mat = NULL,
  r_lab_name = NULL,
  r_trt = NULL,
  r_V = NULL,
  a0 = 1,
  b0 = 1,
  eps = 1e-08,
  knn = 10L,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  NUM_THREADS = 1L,
  IMPUTE_BY_KNN = FALSE
)
}
\arguments{
\item{mtx_file}{data file}

\item{row_file}{row file}

\item{col_file}{column file}

\item{r_cols}{cell (col) names}

\item{r_indv}{membership for the cells (\code{r_cols})}

\item{r_annot}{label annotation for the (\code{r_cols})}

\item{r_annot_mat}{label annotation matrix (cell x type) (default: NULL)}

\item{r_lab_name}{label names (default: everything in \code{r_annot})}

\item{r_trt}{treatment assignment (default: NULL)}

\item{r_V}{SVD factors (default: NULL)}

\item{a0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{b0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{eps}{small number (default: 1e-8)}

\item{knn}{k-NN matching}

\item{KNN_BILINK}{num. of bidirectional links (default: 10)}

\item{KNN_NNLIST}{num. of nearest neighbor lists (default: 10)}

\item{NUM_THREADS}{number of threads for multi-core processing}

\item{IMPUTE_BY_KNN}{imputation by kNN alone (default: FALSE)}
}
\value{
a list of inference results
}
\description{
Create pseudo-bulk data by aggregating columns
}
