% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_aggregate_pairwise}
\alias{rcpp_mmutil_aggregate_pairwise}
\title{Create pseudo-bulk data for pairwise comparisons}
\usage{
rcpp_mmutil_aggregate_pairwise(
  mtx_file,
  row_file,
  col_file,
  r_indv,
  r_V,
  r_cols = NULL,
  r_annot = NULL,
  r_annot_mat = NULL,
  r_lab_name = NULL,
  a0 = 1,
  b0 = 1,
  eps = 1e-08,
  knn_cell = 10L,
  knn_indv = 1L,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  NUM_THREADS = 1L
)
}
\arguments{
\item{mtx_file}{data file}

\item{row_file}{row file}

\item{col_file}{column file}

\item{r_indv}{membership for the cells (\code{r_cols})}

\item{r_V}{SVD factors}

\item{r_cols}{cell (col) names (if we want to take a subset)}

\item{r_annot}{label annotation for the (\code{r_cols})}

\item{r_annot_mat}{label annotation matrix (cell x type) (default: NULL)}

\item{r_lab_name}{label names (default: everything in \code{r_annot})}

\item{a0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{b0}{hyperparameter for gamma(a0, b0) (default: 1)}

\item{eps}{small number (default: 1e-8)}

\item{knn_cell}{k-NN matching between cells}

\item{knn_indv}{k-NN matching between individuals}

\item{KNN_BILINK}{num. of bidirectional links (default: 10)}

\item{KNN_NNLIST}{num. of nearest neighbor lists (default: 10)}

\item{NUM_THREADS}{number of threads for multi-core processing}
}
\value{
a list of inference results
}
\description{
Create pseudo-bulk data for pairwise comparisons
}
\examples{
options(stringsAsFactors = FALSE)
unlink(list.files(pattern = "sim_test"))
.sim <- simulate_indv_glm()
.dat <- rcpp_mmutil_simulate_poisson(.sim$obs.mu,
                                             .sim$rho,
                                             "sim_test")

.annot <- read.table(.dat$indv,
                     col.names = c("col", "ind"))
.annot$trt <- .sim$W[match(.annot$ind, 1:length(.sim$W))]
.annot$ct <- "ct1"

## simple PCA
.pca <- rcpp_mmutil_pca(.dat$mtx, 10)

.agg <- rcpp_mmutil_aggregate_pairwise(mtx_file = .dat$mtx,
                                                row_file = .dat$row,
                                                col_file = .dat$col,
                                                r_indv = .annot$ind,
                                                r_V = .pca$V,
                                                r_annot = .annot$ct,
                                                r_lab_name = "ct1",
                                                knn_cell = 50,
                                                knn_indv = 3)

.names <- lapply(colnames(.agg$delta), strsplit, split="[_]")
.names <- lapply(.names, function(x) unlist(x))
.pairs <- data.frame(do.call(rbind, .names), stringsAsFactors=FALSE)
colnames(.pairs) <- c("src","tgt","ct")

w.src <- .sim$W[as.integer(.pairs$src)]
w.tgt <- .sim$W[as.integer(.pairs$tgt)]
w.delta <- w.src - w.tgt

.agg$delta[.agg$delta < 0] <- NA

par(mfrow=c(2,length(.sim$causal)))
for(k in .sim$causal){
    boxplot(.agg$delta[k, w.delta<0],
            .agg$delta[k, w.delta == 0],
            .agg$delta[k, w.delta>0])
}
non.causal <- setdiff(1:nrow(.sim$obs.mu),
              .sim$causal)
for(k in sample(non.causal,length(.sim$causal))){
    boxplot(.agg$delta[k, w.delta<0],
            .agg$delta[k, w.delta == 0],
            .agg$delta[k, w.delta>0])
}
## clean up temp directory
unlink(list.files(pattern = "sim_test"))

}
