% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_bbknn}
\alias{rcpp_mmutil_bbknn}
\title{BBKNN(Batch-balancing kNN) adjustment of SVD factors}
\usage{
rcpp_mmutil_bbknn(
  svd_v,
  svd_d,
  r_batches,
  knn,
  RECIPROCAL_MATCH = TRUE,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  NUM_THREADS = 1L
)
}
\arguments{
\item{svd_v}{(n x L) n number of data points}

\item{svd_d}{(L x 1) singular values}

\item{r_batches}{batch names (n x 1)}

\item{knn}{kNN parameter k}

\item{RECIPROCAL_MATCH}{do reciprocal match (default: TRUE)}

\item{KNN_BILINK}{num. of bidirectional links (default: 10)}

\item{KNN_NNLIST}{num. of nearest neighbor lists (default: 10)}

\item{NUM_THREADS}{number of threads for multi-core processing}
}
\value{
a list of (1) factors.adjusted (2) D (3) V (4) knn
}
\description{
BBKNN(Batch-balancing kNN) adjustment of SVD factors
}
