% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_bbknn_pca}
\alias{rcpp_mmutil_bbknn_pca}
\title{BBKNN(Batch-balancing kNN)-adjusted PCA}
\usage{
rcpp_mmutil_bbknn_pca(
  mtx_file,
  r_batches,
  knn,
  RANK,
  RECIPROCAL_MATCH = TRUE,
  TAKE_LN = TRUE,
  TAU = 1,
  COL_NORM = 10000,
  EM_ITER = 0L,
  EM_TOL = 1e-04,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  LU_ITER = 5L,
  row_weight_file = "",
  NUM_THREADS = 1L,
  BLOCK_SIZE = 10000L
)
}
\arguments{
\item{mtx_file}{data file (feature x n)}

\item{r_batches}{batch names (n x 1)}

\item{knn}{kNN parameter k}

\item{RANK}{SVD rank}

\item{RECIPROCAL_MATCH}{do reciprocal match (default: TRUE)}

\item{TAKE_LN}{take log(1 + x) trans or not}

\item{TAU}{regularization parameter (default: 1)}

\item{COL_NORM}{column normalization}

\item{EM_ITER}{EM iteration for factorization (default: 0)}

\item{EM_TOL}{EM convergence (default: 1e-4)}

\item{KNN_BILINK}{num. of bidirectional links (default: 10)}

\item{KNN_NNLIST}{num. of nearest neighbor lists (default: 10)}

\item{LU_ITER}{LU iteration}

\item{row_weight_file}{row-wise weight file}

\item{NUM_THREADS}{number of threads for multi-core processing}

\item{BLOCK_SIZE}{number of columns per block}
}
\value{
a list of (1) factors.adjusted (2) U (3) D (4) V
}
\description{
BBKNN(Batch-balancing kNN)-adjusted PCA
}
