% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_network_topic_data}
\alias{rcpp_mmutil_network_topic_data}
\title{Construct a kNN cell-cell interaction network and identify gene topics}
\usage{
rcpp_mmutil_network_topic_data(
  mtx_file,
  row_file,
  col_file,
  latent_factor,
  knn,
  output,
  r_batches = NULL,
  CUTOFF = 0.01,
  WEIGHTED = FALSE,
  MAXW = 1,
  KNN_BILINK = 10L,
  KNN_NNLIST = 10L,
  NUM_THREADS = 1L
)
}
\arguments{
\item{mtx_file}{data file (feature x n)}

\item{row_file}{row file (feature x 1)}

\item{col_file}{row file (n x 1)}

\item{latent_factor}{(n x K)}

\item{knn}{kNN parameter}

\item{output}{a file header for resulting files}

\item{r_batches}{batch names (n x 1, default: NULL)}

\item{CUTOFF}{expression present/absent call cutoff (default: 1e-2)}

\item{KNN_BILINK}{num. of bidirectional links (default: 10)}

\item{KNN_NNLIST}{num. of nearest neighbor lists (default: 10)}

\item{NUM_THREADS}{number of threads for multi-core processing}
}
\value{
feature.incidence, sample.incidence, edges, adjacency matrix files
}
\description{
Construct a kNN cell-cell interaction network and identify gene topics
}
