% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mmutil_simulate_poisson}
\alias{rcpp_mmutil_simulate_poisson}
\title{Simulation Poisson data}
\usage{
rcpp_mmutil_simulate_poisson(mu, rho, output, r_indv = NULL)
}
\arguments{
\item{output}{header for ${output}.{mtx.gz,cols.gz,indv.gz}}

\item{r_indv}{N x 1 individual membership (1-based, [1 .. n])}

\item{Mu}{depth-adjusted mean matrix (M x n), M=#features and n=#indv}

\item{Rho}{column depth vector (N x 1), N=#cells}
}
\value{
a list of file names: {output}.{mtx,rows,cols}.gz
}
\description{
Simulation Poisson data
}
\examples{
rr <- rgamma(20, 1, 1)
mm <- matrix(rgamma(10 * 2, 1, 1), 10, 2)
data.hdr <- "test_sim"
.files <- mmutilR::rcpp_mmutil_simulate_poisson(mm, rr, data.hdr)
Y <- Matrix::readMM(.files$mtx)
print(Y)
A <- read.table(.files$indv, col.names = c("col", "ind"))
head(A)
unlink(list.files(pattern = data.hdr))

}
