% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulate.deg.data}
\alias{simulate.deg.data}
\title{Simulate single-cell MTX data for DEG analysis}
\usage{
\method{simulate}{deg.data}(file.header, ...)
}
\arguments{
\item{file.header}{file set header}

\item{nind}{num of individuals}

\item{ngene}{num of genes/features}

\item{ncausal}{num of causal genes}

\item{ncovar.conf}{num of confounding covariates}

\item{ncovar.batch}{num of confounding batch variables}

\item{ncell.ind}{num of cells per individual}

\item{pve.1}{variance of treatment/disease effect}

\item{pve.c}{variance of confounding effect}

\item{pve.a}{variance of confounders to the assignment}

\item{rho.a}{rho ~ gamma(a, b)}

\item{rho.b}{rho ~ gamma(a, b)}

\item{rseed}{random seed}

\item{exposure.type}{"binary" or "continuous"}
}
\value{
simulation results
}
\description{
Simulate single-cell MTX data for DEG analysis
}
\details{
The simulation result list will have two lists:

1. `data`: (a) a matrix market data file `data$mtx`, (b) a file with row names `data$row`, (c) a file with column names `data$col`, (d) an indexing file for the columns `data$idx`, (e) a mapping file between column and individual names "indv"

2. `indv`: (a) `obs.mu` observed (noisy) gene x individual matrix, (b) `clean.mu` clean gene x individual matrix, (c) `X` confounder x individual matrix (d) `W` individual-level treatment assignment (e) `rho` sequencing depth (f) `causal` causal genes
}
