% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulate_indv_eqtl}
\alias{simulate_indv_eqtl}
\title{Simulate individual-level eQTL data}
\usage{
simulate_indv_eqtl(
  X,
  h2,
  pve.u1.by.x = 0.4,
  pve.y.by.u1 = 0.2,
  pve.y.by.u0 = 0.3,
  n.causal.snps = 1,
  n.causal.genes = 5,
  n.u1 = 3,
  n.u0 = 3,
  n.genes = 50
)
}
\arguments{
\item{X}{genotype matrix (individual x SNPs)}

\item{h2}{heritability (proportion of variance of Y explained by genetic X)}

\item{pve.u1.by.x}{(proportion of variance of U1 explained by X)}

\item{pve.y.by.u1}{(proportion of variance of Y explained by U1)}

\item{pve.y.by.u0}{(proportion of variance of Y explained by U0)}

\item{n.causal.snps}{#(X directly affecting on Y)}

\item{n.causal.genes}{#(Y regulated by X)}

\item{n.u1}{#(U1 variables)}

\item{n.u0}{#(U0 variables)}

\item{n.genes}{total number of genes (Y variables)}
}
\value{
simulation results
}
\description{
Simulate individual-level eQTL data
}
\details{
Possible confounding effect model:
\itemize{
\item X -> U1 -> Y : harmful if U1 was adjusted (epigenetics, gene regulation, trans-effects)
\item U0 -> Y : okay to adjust to boost the power (demographic variables, environments)
}

Another rare possibility (we don't consider here):
\itemize{
\item X -> Y -> Z : harmful if Z was adjusted (downstream phenotypes mediated by genes)
}
}
